+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `AttributeName::from_str_unchecked`.
+- Added `AttributeName::from_cow_unchecked`.
 
 ### Removed
 
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `AttributeName`, `AttributeValue` and `BlockName` now implements `PartialOrd`, `Ord` and `Default`.
 - `Section` and `Block` now implements `PartialOrd`, `Ord` and `Hash`.
-- Renamed `AttributeName::try_new` to `from_str`.
+- Renamed `AttributeName::try_new` to `from_cow`.
 
 ### Fixed